Fixes Executing Callbacks Multiple Times
Enter and exit callbacks were being executed multiple times during window
resize. I think this is because there is a delay before the class is
actually added to the element and that is what we are using to determine if
the breakpoint matches. Instead, we use breakpoint name as a data attribute
on the element and set it true or false.